Fixed 64-Bit issue with DbgTraceWindowMessage, no longer requires casts

+++ litestep/TrayService.cpp
@@ -617,7 +617,7 @@ LRESULT CALLBACK TrayService::WindowTrayProc(HWND hWnd, UINT uMsg,
         default:
             if (uMsg >= WM_USER)
             {
-                DbgTraceWindowMessage("Shell_TrayWnd", uMsg, (unsigned long)wParam, (unsigned long)lParam);
+                DbgTraceWindowMessage("Shell_TrayWnd", uMsg, wParam, lParam);
             }
 
             return DefWindowProc(hWnd, uMsg, wParam, lParam);
